# cortex/APIs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DeepCalm API",
    "description": "Performance-маркетинг автопилот для массажного кабинета",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analyst/analyze/{campaign_id}": {
      "post": {
        "tags": [
          "analyst"
        ],
        "summary": "Analyze Campaign",
        "description": "Анализ кампании через AI Analyst\n\nГенерирует детальный анализ эффективности кампании с рекомендациями.",
        "operationId": "analyze_campaign_api_v1_analyst_analyze__campaign_id__post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AnalysisRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyst/chat": {
      "post": {
        "tags": [
          "analyst"
        ],
        "summary": "Chat With Analyst",
        "description": "Чат с AI аналитиком\n\nЗадавайте вопросы о performance маркетинге и получайте экспертные ответы.",
        "operationId": "chat_with_analyst_api_v1_analyst_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyst/campaign/{campaign_id}/data": {
      "get": {
        "tags": [
          "analyst"
        ],
        "summary": "Get Campaign Data For Analysis",
        "description": "Получить данные кампании для анализа\n\nВозвращает структурированные данные кампании с метриками.",
        "operationId": "get_campaign_data_for_analysis_api_v1_analyst_campaign__campaign_id__data_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/weekly/generate": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Generate Weekly Report",
        "description": "Генерация еженедельного отчета\n\nСоздает детальный отчет с AI анализом за указанный период.",
        "operationId": "generate_weekly_report_api_v1_reports_weekly_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReportGenerationRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/weekly/preview": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Preview Weekly Data",
        "description": "Превью данных для отчета\n\nВозвращает сырые данные без AI анализа для быстрого просмотра.",
        "operationId": "preview_weekly_data_api_v1_reports_weekly_preview_get",
        "parameters": [
          {
            "name": "weeks_back",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Weeks Back"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/weekly/email": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Send Weekly Report Email",
        "description": "Отправка отчета по email\n\nГенерирует отчет и отправляет на email из настроек.",
        "operationId": "send_weekly_report_email_api_v1_reports_weekly_email_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReportGenerationRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisRequest": {
        "properties": {
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question",
            "description": "Дополнительный вопрос для анализа"
          }
        },
        "type": "object",
        "title": "AnalysisRequest",
        "description": "Схема запроса анализа кампании"
      },
...
      "CampaignAnalysisResponse": {
        "properties": {
          "campaign_id": {
            "type": "integer",
            "title": "Campaign Id",
            "description": "ID кампании"
          },
          "analysis": {
            "type": "string",
            "title": "Analysis",
            "description": "Текст анализа от AI"
          },
          "metrics": {
            "$ref": "#/components/schemas/app__schemas__analyst__CampaignMetrics"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations",
            "description": "Извлеченные рекомендации"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At",
            "description": "Время генерации анализа"
          },
          "token_usage": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TokenUsage"
              }
            ],
            "description": "Использование токенов"
          }
        },
        "type": "object",
        "required": [
          "campaign_id",
          "analysis",
          "metrics",
          "recommendations",
          "generated_at",
          "token_usage"
        ],
        "title": "CampaignAnalysisResponse",
        "description": "Схема ответа анализа кампании"
      },
...
      "CampaignMetric": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "ID кампании"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Название кампании"
          },
          "sku": {
            "type": "string",
            "title": "Sku",
            "description": "SKU продукта"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Статус кампании"
          },
          "leads": {
            "type": "integer",
            "title": "Leads",
            "description": "Количество лидов"
          },
          "conversions": {
            "type": "integer",
            "title": "Conversions",
            "description": "Количество конверсий"
          },
          "revenue": {
            "type": "number",
            "title": "Revenue",
            "description": "Выручка (рублей)"
          },
          "spend": {
            "type": "number",
            "title": "Spend",
            "description": "Расход (рублей)"
          },
          "roas": {
            "type": "number",
            "title": "Roas",
            "description": "Return on Ad Spend"
          },
          "cac": {
            "type": "number",
            "title": "Cac",
            "description": "Customer Acquisition Cost"
          },
          "target_cac": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Cac",
            "description": "Целевой CAC"
          },
          "target_roas": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Roas",
            "description": "Целевой ROAS"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "sku",
          "status",
          "leads",
          "conversions",
          "revenue",
          "spend",
          "roas",
          "cac"
        ],
        "title": "CampaignMetric",
        "description": "Метрики кампании"
      },
...
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 1000,
            "title": "Message",
            "description": "Сообщение аналитику"
          },
          "campaign_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Id",
            "description": "ID кампании для контекста"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "description": "Схема запроса чата с аналитиком"
      },
      "ChatResponse": {
        "properties": {
          "response": {
            "type": "string",
            "title": "Response",
            "description": "Ответ аналитика"
          },
          "campaign_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Id",
            "description": "ID кампании из контекста"
          }
        },
        "type": "object",
        "required": [
          "response"
        ],
        "title": "ChatResponse",
        "description": "Схема ответа чата"
      },
...
      "DetailedReportData": {
        "properties": {
          "top_performers": {
            "items": {
              "$ref": "#/components/schemas/CampaignMetric"
            },
            "type": "array",
            "title": "Top Performers",
            "description": "Топ-кампании"
          },
          "needs_attention": {
            "items": {
              "$ref": "#/components/schemas/CampaignMetric"
            },
            "type": "array",
            "title": "Needs Attention",
            "description": "Кампании, требующие внимания"
          },
          "all_campaigns": {
            "items": {
              "$ref": "#/components/schemas/CampaignMetric"
            },
            "type": "array",
            "title": "All Campaigns",
            "description": "Все кампании"
          }
        },
        "type": "object",
        "required": [
          "top_performers",
          "needs_attention",
          "all_campaigns"
        ],
        "title": "DetailedReportData",
        "description": "Детальные данные отчета"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportGenerationRequest": {
        "properties": {
          "weeks_back": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Weeks Back",
            "description": "Количество недель назад (1-12)",
            "default": 1
          }
        },
        "type": "object",
        "title": "ReportGenerationRequest",
        "description": "Схема запроса генерации отчета"
      },
      "ReportPeriod": {
        "properties": {
          "start_date": {
            "type": "string",
            "title": "Start Date",
            "description": "Дата начала (ISO формат)"
          },
          "end_date": {
            "type": "string",
            "title": "End Date",
            "description": "Дата окончания (ISO формат)"
          },
          "weeks": {
            "type": "integer",
            "title": "Weeks",
            "description": "Количество недель"
          }
        },
        "type": "object",
        "required": [
          "start_date",
          "end_date",
          "weeks"
        ],
        "title": "ReportPeriod",
        "description": "Период отчета"
      },
      "ReportSettings": {
        "properties": {
          "reports_email": {
            "type": "string",
            "title": "Reports Email",
            "description": "Email для отправки отчетов"
          },
          "auto_generated": {
            "type": "boolean",
            "title": "Auto Generated",
            "description": "Автоматически сгенерирован"
          }
        },
        "type": "object",
        "required": [
          "reports_email",
          "auto_generated"
        ],
        "title": "ReportSettings",
        "description": "Настройки отчета"
      },
...
      "ReportSummary": {
        "properties": {
          "total_leads": {
            "type": "integer",
            "title": "Total Leads",
            "description": "Общее количество лидов"
          },
          "total_conversions": {
            "type": "integer",
            "title": "Total Conversions",
            "description": "Общее количество конверсий"
          },
          "total_revenue": {
            "type": "number",
            "title": "Total Revenue",
            "description": "Общая выручка (рублей)"
          },
          "conversion_rate": {
            "type": "number",
            "title": "Conversion Rate",
            "description": "Общий процент конверсии"
          },
          "active_campaigns": {
            "type": "integer",
            "title": "Active Campaigns",
            "description": "Количество активных кампаний"
          }
        },
        "type": "object",
        "required": [
          "total_leads",
          "total_conversions",
          "total_revenue",
          "conversion_rate",
          "active_campaigns"
        ],
        "title": "ReportSummary",
        "description": "Сводка метрик"
      },
...
      "TokenUsage": {
        "properties": {
          "prompt_tokens": {
            "type": "integer",
            "title": "Prompt Tokens",
            "description": "Токены в промпте"
          },
          "completion_tokens": {
            "type": "integer",
            "title": "Completion Tokens",
            "description": "Токены в ответе"
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens",
            "description": "Общее количество токенов"
          }
        },
        "type": "object",
        "required": [
          "prompt_tokens",
          "completion_tokens",
          "total_tokens"
        ],
        "title": "TokenUsage",
        "description": "Использование токенов OpenAI"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeeklyReportResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "ID отчета"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At",
            "description": "Время генерации (ISO формат)"
          },
          "period": {
            "$ref": "#/components/schemas/ReportPeriod"
          },
          "summary": {
            "$ref": "#/components/schemas/ReportSummary"
          },
          "ai_analysis": {
            "type": "string",
            "title": "Ai Analysis",
            "description": "AI анализ отчета"
          },
          "detailed_data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DetailedReportData"
              }
            ],
            "description": "Детальные данные"
          },
          "settings": {
            "$ref": "#/components/schemas/ReportSettings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "generated_at",
          "period",
          "summary",
          "ai_analysis",
          "detailed_data",
          "settings"
        ],
        "title": "WeeklyReportResponse",
        "description": "Схема ответа еженедельного отчета"
      },
      "app__schemas__analyst__CampaignMetrics": {
        "properties": {
          "total_leads": {
            "type": "integer",
            "title": "Total Leads",
            "description": "Общее количество лидов"
          },
          "total_conversions": {
            "type": "integer",
            "title": "Total Conversions",
            "description": "Общее количество конверсий"
          },
          "conversion_rate": {
            "type": "number",
            "title": "Conversion Rate",
            "description": "Процент конверсии"
          },
          "total_revenue": {
            "type": "number",
            "title": "Total Revenue",
            "description": "Общая выручка (рублей)"
          },
          "total_spend": {
            "type": "number",
            "title": "Total Spend",
            "description": "Общие расходы (рублей)"
          },
          "roas": {
            "type": "number",
            "title": "Roas",
            "description": "Return on Ad Spend"
          },
          "cac": {
            "type": "number",
            "title": "Cac",
            "description": "Customer Acquisition Cost (рублей)"
          },
          "period_days": {
            "type": "integer",
            "title": "Period Days",
            "description": "Период анализа (дней)"
          }
        },
        "type": "object",
        "required": [
          "total_leads",
          "total_conversions",
          "conversion_rate",
          "total_revenue",
          "total_spend",
          "roas",
          "cac",
          "period_days"
        ],
        "title": "CampaignMetrics",
        "description": "Метрики кампании"
      },
...
    }
  }
}